refactored scroll form so to easier hide section
as a consequence fixed issue:

fixes: #2686
